+++ Assets/Scripts/MVC/Model/RoomModel.cs
@@ -5,7 +5,7 @@ namespace MVC.Model
 {
     public class RoomModel : BaseModel<RoomModel>
     {
-        public short RoomID { get; private set; } // 房间ID
+        public int RoomID { get; private set; } // 房间ID
         public short CurrentCycle { get; private set; } // 当前圈数
         public short TotalCycle { get; private set; } // 总圈数
         public byte DealerWind { get; private set; } // 门风
